update trả hàng 15:44 18/06/24

+++ WebAccountant/Repository/Implement/KtscRepo.cs
@@ -332,7 +332,7 @@ namespace WebAccountant.Repository.Implement
             return returnForm;
         }
 
-        public async Task<bool> RefundPackageSell(IEnumerable<int> sttSc)
+        public async Task<bool> RefundPackageSell(IEnumerable<double> sttSc)
         {
             foreach (var item in sttSc)
             {
